Close 3rd-review findings: Helm guard, exports robustness, peer-dep hygiene
Helm chart (deploy/helm/radar):
- Add `fail` guard when hub.enabled=true but neither hub.token nor
  hub.existingSecret is set. Previously installed silently to CrashLoopBackOff.
- New CI job lints + renders the chart in four configurations (default,
  hub+inline token, hub+existingSecret, hub-missing-token→must-fail) so
  chart drift can't slip past review.

k8s-ui exports (packages/k8s-ui/package.json):
- Add `./components/resources/renderers/*` — the existing
  `./components/resources/*` pattern doesn't cover it, and 100+ Radar
  imports depended on the Vite alias in web/ to paper over the gap.
  Strict-resolver consumers (hub installing via npm) would have hit
  ERR_PACKAGE_PATH_NOT_EXPORTED.
- Array form for every mixed-extension directory
  (./components/{resources,ui,gitops,timeline}/*, ./hooks/*) — hooks/
  has both useAnimatedUnmount.ts and useKeyboardShortcuts.tsx; gitops/,
  timeline/, ui/ all mix .tsx files with an index.ts. Previous single-
  extension patterns silently failed on the other half.

React-query-meta augmentation (web/src/):
- Inline the `declare module` into RadarApp.tsx directly instead of a
  separate side-effect-imported file. web/package.json declares
  sideEffects: ["*.css"], which risks the side-effect import being
  tree-shaken by aggressive bundlers. Inlining makes the augmentation
  impossible to drop as long as the consumer imports RadarApp.

web/package.json peer/dep hygiene:
- @xyflow/react and elkjs move deps → peerDeps (React-context libs
  need to be deduped with k8s-ui's copy).
- @xterm/* removed from deps entirely — not imported from web/src,
  only transitively via k8s-ui which already peers them.
- @tailwindcss/typography moves deps → devDeps (build-time only;
  consumers bring their own Tailwind config).
- lucide-react peer tightens from "*" to ">=0.400.0" to match k8s-ui.
- @skyhook-io/k8s-ui peer bumps to >=1.5.0 to enforce the exports
  update. Publishing radar-app against <1.5.0 k8s-ui would install a
  broken combination.
- Export getAuthHeaders + getCredentialsMode for getter/setter symmetry.

Publish workflow (.github/workflows/publish-radar-app.yml):
- Header comment calling out the publish ordering: k8s-ui 1.5.0+ must
  ship first or radar-app install fails peer-dep resolution.

Other:
- Add web/README.md so npm doesn't show a blank description tab.
