+++ .github/workflows/publish-radar-app.yml
@@ -1,5 +1,11 @@
 name: Publish @skyhook-io/radar-app
 
+# Radar's source imports deep paths from @skyhook-io/k8s-ui
+# (./types/*, ./utils/*, ./components/gitops/*, ./components/timeline/*,
+# ./components/resources/renderers/*). Those subpath export patterns were
+# added in k8s-ui v1.5.0. Publish a matching k8s-ui version FIRST before
+# tagging a radar-app release — peer dep resolves to '>=1.5.0'.
+
 on:
   push:
     tags:
