+++ deploy/helm/radar/templates/deployment.yaml
@@ -87,6 +87,9 @@ spec:
             {{- end }}
             {{- end }}
             {{- if .Values.hub.enabled }}
+            {{- if not (or .Values.hub.token .Values.hub.existingSecret) }}
+            {{- fail "hub.token or hub.existingSecret must be set when hub.enabled=true" }}
+            {{- end }}
             - --hub-url={{ required "hub.url is required when hub.enabled=true" .Values.hub.url }}
             - --cluster-name={{ required "hub.clusterName is required when hub.enabled=true" .Values.hub.clusterName }}
             {{- /* hub.token is injected via RADAR_HUB_TOKEN env var below, not the command line */}}
