+++ web/package.json
@@ -30,13 +30,7 @@
     "@fontsource-variable/dm-sans": "^5.2.8",
     "@fontsource/dm-mono": "^5.2.7",
     "@monaco-editor/react": "^4.7.0",
-    "@tailwindcss/typography": "^0.5.19",
-    "@xterm/addon-fit": "^0.11.0",
-    "@xterm/addon-web-links": "^0.12.0",
-    "@xterm/xterm": "^6.0.0",
-    "@xyflow/react": "^12.10.2",
     "diff": "^9.0.0",
-    "elkjs": "^0.11.1",
     "react-markdown": "^10.1.0",
     "react-virtuoso": "^4.18.5",
     "remark-gfm": "^4.0.1",
@@ -44,18 +38,21 @@
     "yaml": "^2.8.3"
   },
   "peerDependencies": {
-    "@skyhook-io/k8s-ui": ">=1.4.0",
+    "@skyhook-io/k8s-ui": ">=1.5.0",
     "@tanstack/react-query": ">=5",
+    "@xyflow/react": ">=12.0.0",
+    "clsx": ">=2",
+    "elkjs": ">=0.9.0",
+    "lucide-react": ">=0.400.0",
     "react": ">=19",
     "react-dom": ">=19",
     "react-router-dom": ">=7",
-    "clsx": ">=2",
-    "lucide-react": "*",
     "tailwind-merge": ">=2"
   },
   "devDependencies": {
     "@playwright/test": "^1.59.1",
     "@skyhook-io/k8s-ui": "*",
+    "@tailwindcss/typography": "^0.5.19",
     "@tailwindcss/vite": "^4.2.1",
     "@tanstack/react-query": "^5.99.0",
     "@types/diff": "^8.0.0",
@@ -63,7 +60,9 @@
     "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.3",
     "@vitejs/plugin-react": "^6.0.1",
+    "@xyflow/react": "^12.10.2",
     "clsx": "^2.1.1",
+    "elkjs": "^0.11.1",
     "lucide-react": "^1.8.0",
     "postcss": "^8.5.8",
     "prettier": "^3.8.1",
