+++ web/src/index.ts
@@ -1,13 +1,8 @@
 // @skyhook-io/radar-app — Radar's full web UI as a reusable React component.
 //
-// Source-only distribution (same model as @skyhook-io/k8s-ui): consumers
-// are Vite/Next.js projects that transpile TS/TSX and resolve aliases at
-// build time. No bundle step.
-//
-// Radar's own binary entry (main.tsx) renders the same RadarApp component,
-// so standalone and embedded modes share exactly one code path.
-import './react-query-meta';
-
+// Source-only package (main points at .ts, no dist/). Consumers need a
+// bundler that transpiles TSX and resolves workspace-style peer deps. The
+// same source is consumed by Radar's binary via main.tsx.
 export { RadarApp, type RadarAppProps } from './RadarApp';
 export {
   setApiBase,
@@ -16,5 +11,7 @@ export {
   setCredentialsMode,
   getApiBase,
   getBasename,
+  getAuthHeaders,
+  getCredentialsMode,
 } from './api/config';
 export type { NavCustomization } from './context/NavCustomization';
